Move package to `pkg/utils/gardener/secretsrotation`
